feat(cli): add --clean-target option for granular clean/flush/anonymize control
- Add --clean-target (-ct) option with choices: all, model, predict
- Allow users to clean only prediction dir while keeping model intact
- Update Cleaner, Flusher, Anonymizer classes to support target parameter
- Add file existence checks to prevent errors on missing files
- Update README with new options and usage examples

Closes ersilia-os/zairachem-docker#36

+++ zairachem/finish/finish.py
@@ -19,52 +19,87 @@ from zairachem.base.vars import (
 from zairachem.base import ZairaBase
 from zairachem.base.utils.pipeline import PipelineStep
 
+CLEAN_TARGET_ALL = "all"
+CLEAN_TARGET_MODEL = "model"
+CLEAN_TARGET_PREDICT = "predict"
+VALID_CLEAN_TARGETS = [CLEAN_TARGET_ALL, CLEAN_TARGET_MODEL, CLEAN_TARGET_PREDICT]
+
 
 class Cleaner(ZairaBase):
-  def __init__(self, path):
+  def __init__(self, path, target=CLEAN_TARGET_ALL):
     ZairaBase.__init__(self)
     if path is None:
       self.path = self.get_output_dir()
     else:
       self.path = path
     self.output_dir = os.path.abspath(self.path)
+    self.target = target
+    self._is_predict = self.is_predict()
+    if self._is_predict:
+      self.trained_dir = self.get_trained_dir()
+    else:
+      self.trained_dir = self.path
     assert os.path.exists(self.output_dir)
 
   def _clean_descriptors_by_subfolder(self, path, subfolder):
-    path = os.path.join(path, subfolder)
-    for d in os.listdir(path):
+    full_path = os.path.join(path, subfolder)
+    if not os.path.exists(full_path):
+      return
+    for d in os.listdir(full_path):
       if d.startswith("fp2sim"):
         continue
-      if os.path.isdir(os.path.join(path, d)):
-        self._clean_descriptors_by_subfolder(path, d)
+      if os.path.isdir(os.path.join(full_path, d)):
+        self._clean_descriptors_by_subfolder(full_path, d)
       else:
         if d.endswith(".h5"):
-          os.remove(os.path.join(path, d))
+          os.remove(os.path.join(full_path, d))
 
-  def _clean_descriptors(self):
-    self._clean_descriptors_by_subfolder(self.path, DESCRIPTORS_SUBFOLDER)
+  def _clean_descriptors(self, path):
+    self._clean_descriptors_by_subfolder(path, DESCRIPTORS_SUBFOLDER)
 
   def run(self):
-    self.logger.debug("Cleaning descriptors by subfolder")
-    self._clean_descriptors()
+    if self.target == CLEAN_TARGET_ALL:
+      self.logger.debug("Cleaning descriptors from output directory")
+      self._clean_descriptors(self.output_dir)
+      if self._is_predict and self.trained_dir != self.output_dir:
+        self.logger.debug("Cleaning descriptors from model directory")
+        self._clean_descriptors(self.trained_dir)
+    elif self.target == CLEAN_TARGET_MODEL:
+      if self._is_predict:
+        self.logger.debug("Cleaning descriptors from model directory only")
+        self._clean_descriptors(self.trained_dir)
+      else:
+        self.logger.debug("Cleaning descriptors from model directory (fit mode)")
+        self._clean_descriptors(self.output_dir)
+    elif self.target == CLEAN_TARGET_PREDICT:
+      if self._is_predict:
+        self.logger.debug("Cleaning descriptors from predict directory only")
+        self._clean_descriptors(self.output_dir)
+      else:
+        self.logger.warning("Cannot clean predict directory during fit - skipping")
 
 
 class Flusher(ZairaBase):
-  def __init__(self, path):
+  def __init__(self, path, target=CLEAN_TARGET_ALL):
     ZairaBase.__init__(self)
     if path is None:
       self.path = self.get_output_dir()
     else:
       self.path = path
     self.output_dir = os.path.abspath(self.path)
-    self.trained_dir = self.get_trained_dir()
+    self.target = target
+    self._is_predict = self.is_predict()
+    if self._is_predict:
+      self.trained_dir = self.get_trained_dir()
+    else:
+      self.trained_dir = self.path
     assert os.path.exists(self.output_dir)
-    assert os.path.exists(self.trained_dir)
 
   def _remover(self, path):
+    if not os.path.exists(path):
+      return
     rm_dirs = []
     rm_files = []
-
     for root, dirs, files in os.walk(path):
       for filename in files:
         if filename.endswith(".json") or filename.endswith(".csv"):
@@ -83,23 +118,45 @@ class Flusher(ZairaBase):
         shutil.rmtree(d)
 
   def _flush(self, path):
-    self.logger.debug("Removing files descriptors folder in {0}".format(path))
+    self.logger.debug("Removing files from descriptors folder in {0}".format(path))
     self._remover(os.path.join(path, DESCRIPTORS_SUBFOLDER))
     self.logger.debug("Removing files from estimators folder in {0}".format(path))
     self._remover(os.path.join(path, ESTIMATORS_SUBFOLDER))
 
   def run(self):
-    self._flush(self.output_dir)
-    self._flush(self.trained_dir)
+    if self.target == CLEAN_TARGET_ALL:
+      self._flush(self.output_dir)
+      if self._is_predict and self.trained_dir != self.output_dir:
+        self._flush(self.trained_dir)
+    elif self.target == CLEAN_TARGET_MODEL:
+      if self._is_predict:
+        self.logger.debug("Flushing model directory only")
+        self._flush(self.trained_dir)
+      else:
+        self.logger.debug("Flushing model directory (fit mode)")
+        self._flush(self.output_dir)
+    elif self.target == CLEAN_TARGET_PREDICT:
+      if self._is_predict:
+        self.logger.debug("Flushing predict directory only")
+        self._flush(self.output_dir)
+      else:
+        self.logger.warning("Cannot flush predict directory during fit - skipping")
 
 
 class Anonymizer(ZairaBase):
-  def __init__(self, path):
+  def __init__(self, path, target=CLEAN_TARGET_ALL):
     ZairaBase.__init__(self)
     if path is None:
       self.path = self.get_output_dir()
     else:
       self.path = path
+    self.output_dir = os.path.abspath(self.path)
+    self.target = target
+    self._is_predict = self.is_predict()
+    if self._is_predict:
+      self.trained_dir = self.get_trained_dir()
+    else:
+      self.trained_dir = self.path
     self._empty_string = "NA"
 
   def _remove_file_if_exists(self, file_name):
@@ -107,6 +164,8 @@ class Anonymizer(ZairaBase):
       os.remove(file_name)
 
   def _replace_sensitive_columns(self, file_name):
+    if not os.path.exists(file_name):
+      return
     columns = ["input-smiles", "smiles", "inchikey"]
     df = pd.read_csv(file_name)
     df_columns = set(list(df.columns))
@@ -116,6 +175,8 @@ class Anonymizer(ZairaBase):
     df.to_csv(file_name, index=False)
 
   def _remove_sensitive_columns(self, file_name):
+    if not os.path.exists(file_name):
+      return
     desc_columns = [
       "nHBD",
       "nHBA",
@@ -146,32 +207,33 @@ class Anonymizer(ZairaBase):
     df.to_csv(file_name, index=False)
 
   def _replace_first_last_descriptors(self, file_name):
+    if not os.path.exists(file_name):
+      return
     df = pd.read_csv(file_name)
     df.loc[0:1, df.columns[1:]] = self._empty_string
     df.to_csv(file_name, index=False)
 
-  def _remove_raw_input(self):
-    file_name = os.path.join(self.path, RAW_INPUT_FILENAME + ".csv")
+  def _remove_raw_input(self, path):
+    file_name = os.path.join(path, RAW_INPUT_FILENAME + ".csv")
     self._remove_file_if_exists(file_name)
 
-  def _remove_output_table_xlsx(self):
-    file_name = os.path.join(self.path, OUTPUT_XLSX_FILENAME)
+  def _remove_output_table_xlsx(self, path):
+    file_name = os.path.join(path, OUTPUT_XLSX_FILENAME)
     self._remove_file_if_exists(file_name)
 
-  def _remove_applicability_files(self):
-    for fn in os.listdir(os.path.join(self.path, APPLICABILITY_SUBFOLDER)):
-      fn = os.path.join(self.path, APPLICABILITY_SUBFOLDER, fn)
-      self._remove_file_if_exists(fn)
+  def _clear_all_sensitive_columns(self, path):
+    self._replace_sensitive_columns(os.path.join(path, DATA_SUBFOLDER, DATA_FILENAME))
+    self._replace_sensitive_columns(os.path.join(path, OUTPUT_FILENAME))
+    self._remove_sensitive_columns(os.path.join(path, OUTPUT_TABLE_FILENAME))
+    self._replace_sensitive_columns(os.path.join(path, OUTPUT_TABLE_FILENAME))
 
-  def _clear_all_sensitive_columns(self):
-    self._replace_sensitive_columns(os.path.join(self.path, DATA_SUBFOLDER, DATA_FILENAME))
-    self._replace_sensitive_columns(os.path.join(self.path, OUTPUT_FILENAME))
-    self._remove_sensitive_columns(os.path.join(self.path, OUTPUT_TABLE_FILENAME))
-    self._replace_sensitive_columns(os.path.join(self.path, OUTPUT_TABLE_FILENAME))
-
-  def _clear_descriptors(self):
-    Cleaner(path=self.path).run()
-    subfolder = os.path.join(self.path, DESCRIPTORS_SUBFOLDER)
+  def _clear_descriptors(self, path):
+    Cleaner(
+      path=path, target=CLEAN_TARGET_PREDICT if path == self.output_dir else CLEAN_TARGET_MODEL
+    ).run()
+    subfolder = os.path.join(path, DESCRIPTORS_SUBFOLDER)
+    if not os.path.exists(subfolder):
+      return
     for d in os.listdir(subfolder):
       if d.startswith("fp2sim"):
         self._remove_file_if_exists(os.path.join(subfolder, d))
@@ -180,40 +242,60 @@ class Anonymizer(ZairaBase):
       if file_name == "raw_summary.csv":
         self._replace_first_last_descriptors(file_path)
 
-  def _clear_estimators(self):
-    subfolder = os.path.join(self.path, ESTIMATORS_SUBFOLDER).rstrip("/")
+  def _clear_estimators(self, path):
+    subfolder = os.path.join(path, ESTIMATORS_SUBFOLDER).rstrip("/")
+    if not os.path.exists(subfolder):
+      return
     for file_path in glob.iglob(subfolder + "/**", recursive=True):
       file_name = file_path.split("/")[-1]
       if file_name == RESULTS_MAPPED_FILENAME or file_name == RESULTS_UNMAPPED_FILENAME:
         self._replace_sensitive_columns(file_path)
-    self._remove_file_if_exists(os.path.join(subfolder, "manifolds", "data.csv"))
+    manifolds_data = os.path.join(subfolder, "manifolds", "data.csv")
+    self._remove_file_if_exists(manifolds_data)
 
-  def _clear_pool(self):
-    self._remove_sensitive_columns(os.path.join(self.path, POOL_SUBFOLDER, "data.csv"))
-    self._replace_sensitive_columns(
-      os.path.join(self.path, POOL_SUBFOLDER, RESULTS_MAPPED_FILENAME)
-    )
-    self._replace_sensitive_columns(
-      os.path.join(self.path, POOL_SUBFOLDER, RESULTS_UNMAPPED_FILENAME)
-    )
+  def _clear_pool(self, path):
+    pool_data = os.path.join(path, POOL_SUBFOLDER, "data.csv")
+    if os.path.exists(pool_data):
+      self._remove_sensitive_columns(pool_data)
+    self._replace_sensitive_columns(os.path.join(path, POOL_SUBFOLDER, RESULTS_MAPPED_FILENAME))
+    self._replace_sensitive_columns(os.path.join(path, POOL_SUBFOLDER, RESULTS_UNMAPPED_FILENAME))
 
-  def _clear_report(self):
-    self._remove_file_if_exists(
-      os.path.join(self.path, REPORT_SUBFOLDER, "tanimoto-similarity-to-train.png")
-    )
-    self._remove_file_if_exists(os.path.join(self.path, REPORT_SUBFOLDER, "projection-pca.png"))
-    self._remove_file_if_exists(os.path.join(self.path, REPORT_SUBFOLDER, "projection-umap.png"))
-    self._remove_file_if_exists(os.path.join(self.path, REPORT_SUBFOLDER, OUTPUT_TABLE_FILENAME))
+  def _clear_report(self, path):
+    report_folder = os.path.join(path, REPORT_SUBFOLDER)
+    if not os.path.exists(report_folder):
+      return
+    self._remove_file_if_exists(os.path.join(report_folder, "tanimoto-similarity-to-train.png"))
+    self._remove_file_if_exists(os.path.join(report_folder, "projection-pca.png"))
+    self._remove_file_if_exists(os.path.join(report_folder, "projection-umap.png"))
+    self._remove_file_if_exists(os.path.join(report_folder, OUTPUT_TABLE_FILENAME))
+
+  def _anonymize_path(self, path):
+    self._remove_raw_input(path)
+    self._remove_output_table_xlsx(path)
+    self._clear_all_sensitive_columns(path)
+    self._clear_descriptors(path)
+    self._clear_estimators(path)
+    self._clear_pool(path)
+    self._clear_report(path)
 
   def run(self):
-    self._remove_raw_input()
-    self._remove_output_table_xlsx()
-    self._remove_applicability_files()
-    self._clear_all_sensitive_columns()
-    self._clear_descriptors()
-    self._clear_estimators()
-    self._clear_pool()
-    self._clear_report()
+    if self.target == CLEAN_TARGET_ALL:
+      self._anonymize_path(self.output_dir)
+      if self._is_predict and self.trained_dir != self.output_dir:
+        self._anonymize_path(self.trained_dir)
+    elif self.target == CLEAN_TARGET_MODEL:
+      if self._is_predict:
+        self.logger.debug("Anonymizing model directory only")
+        self._anonymize_path(self.trained_dir)
+      else:
+        self.logger.debug("Anonymizing model directory (fit mode)")
+        self._anonymize_path(self.output_dir)
+    elif self.target == CLEAN_TARGET_PREDICT:
+      if self._is_predict:
+        self.logger.debug("Anonymizing predict directory only")
+        self._anonymize_path(self.output_dir)
+      else:
+        self.logger.warning("Cannot anonymize predict directory during fit - skipping")
 
 
 class OutputToExcel(ZairaBase):
@@ -228,6 +310,9 @@ class OutputToExcel(ZairaBase):
     self.output_xlsx = os.path.join(self.path, OUTPUT_XLSX_FILENAME)
 
   def run(self):
+    if not os.path.exists(self.output_csv):
+      self.logger.warning(f"Output table not found: {self.output_csv}")
+      return
     df_o = pd.read_csv(self.output_csv)
     if not os.path.exists(self.performance_csv):
       df_p = None
@@ -240,7 +325,9 @@ class OutputToExcel(ZairaBase):
 
 
 class Finisher(ZairaBase):
-  def __init__(self, path, clean=False, flush=False, anonymize=False):
+  def __init__(
+    self, path, clean=False, flush=False, anonymize=False, clean_target=CLEAN_TARGET_ALL
+  ):
     ZairaBase.__init__(self)
     if path is None:
       self.path = self.get_output_dir()
@@ -249,27 +336,28 @@ class Finisher(ZairaBase):
     self.clean = clean
     self.flush = flush
     self.anonymize = anonymize
+    self.clean_target = clean_target if clean_target in VALID_CLEAN_TARGETS else CLEAN_TARGET_ALL
 
   def _clean_descriptors(self):
-    Cleaner(path=self.path).run()
+    Cleaner(path=self.path, target=self.clean_target).run()
 
   def _flush(self):
-    Flusher(path=self.path).run()
+    Flusher(path=self.path, target=self.clean_target).run()
 
   def _anonymize(self):
-    Anonymizer(path=self.path).run()
+    Anonymizer(path=self.path, target=self.clean_target).run()
 
   def _predictions_file(self):
-    shutil.copy(
-      os.path.join(self.path, POOL_SUBFOLDER, RESULTS_MAPPED_FILENAME),
-      os.path.join(self.path, OUTPUT_FILENAME),
-    )
+    src = os.path.join(self.path, POOL_SUBFOLDER, RESULTS_MAPPED_FILENAME)
+    dst = os.path.join(self.path, OUTPUT_FILENAME)
+    if os.path.exists(src):
+      shutil.copy(src, dst)
 
   def _output_table_file(self):
-    shutil.copy(
-      os.path.join(self.path, REPORT_SUBFOLDER, OUTPUT_TABLE_FILENAME),
-      os.path.join(self.path, OUTPUT_TABLE_FILENAME),
-    )
+    src = os.path.join(self.path, REPORT_SUBFOLDER, OUTPUT_TABLE_FILENAME)
+    dst = os.path.join(self.path, OUTPUT_TABLE_FILENAME)
+    if os.path.exists(src):
+      shutil.copy(src, dst)
 
   def _performance_table_file(self):
     filename = os.path.join(self.path, REPORT_SUBFOLDER, PERFORMANCE_TABLE_FILENAME)
@@ -290,10 +378,13 @@ class Finisher(ZairaBase):
     self._performance_table_file()
     self._to_excel()
     if self.clean:
+      self.logger.info(f"Cleaning with target: {self.clean_target}")
       self._clean_descriptors()
     if self.flush:
+      self.logger.info(f"Flushing with target: {self.clean_target}")
       self._flush()
     if self.anonymize:
+      self.logger.info(f"Anonymizing with target: {self.clean_target}")
       self._anonymize()
 
   def run(self):
@@ -305,5 +396,4 @@ class Finisher(ZairaBase):
       self.logger.warning(
         "[yellow]Finishing setup for requested inferece is already done. Skippign this step![/]"
       )
-
     self.logger.info("[green]All zairachem successfully completed![/]")
